+++ source/network/NetStats.cpp
@@ -72,7 +72,7 @@ const std::vector<ProfileColumn>& CNetStatsTable::GetColumns()
 	if (m_Host)
 	{
 		for (size_t i = 0; i < m_Host->peerCount; ++i)
-			m_ColumnDescriptions.push_back(ProfileColumn("Peer "+CStr(i), 80));
+			m_ColumnDescriptions.push_back(ProfileColumn("Peer "+CStr((unsigned long)i), 80));
 	}
 	else if (m_Peer)
 	{
